# Disable attention mask during new token generation